+++ packages/sdk/src/types/attributes/IQLQuery.ts
@@ -1,0 +1,6 @@
+export interface IQLQuery {
+    "@type"?: "IQLQuery";
+    queryString: string;
+    validFrom?: string;
+    validTo?: string;
+}