+++ packages/sdk/src/types/attributes/requests/ExecuteIQLQueryRequest.ts
@@ -1,0 +1,5 @@
+import { IQLQuery } from "../IQLQuery";
+
+export interface ExecuteIQLQueryRequest {
+    query: IQLQuery;
+}